rename nexus-maybe to nexusMaybe
