+++ src/shared/views/ResourceView.tsx
@@ -21,7 +21,7 @@ import ResourceEditor from '../components/Resources/ResourceEditor';
 import HistoryContainer from '../containers/HistoryContainer';
 import ResourceLinksContainer from '../containers/ResourceLinks';
 import ResourceActionsContainer from '../containers/ResourceActions';
-import { isDeprecated } from '../utils/nexus-maybe';
+import { isDeprecated } from '../utils/nexusMaybe';
 
 const TabPane = Tabs.TabPane;
 const DEFAULT_ACTIVE_TAB_KEY = '#JSON';
